% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choleradeaths_data.R
\docType{data}
\name{choleradeaths}
\alias{choleradeaths}
\title{choleradeaths}
\format{
A data frame with 250 rows and 2 variables:
\describe{
\item{\code{ID}}{integer ID}
\item{\code{Cholera.Deaths}}{integer Number of cholera deaths}
\item{\code{geometry}}{sfc_POINT geometry}
}
}
\source{
\href{https://blog.rtwilson.com/john-snows-famous-cholera-analysis-data-in-modern-gis-formats/}{Robin's Blog}
}
\usage{
choleradeaths
}
\description{
An sf dataset on cholera deaths in Soho from John Snow's analysis adapted from Robin Wilson's digitised dataset.
}
\details{
Data on Cholera Deaths in Soho
}
\keyword{datasets}
