% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcircles.R
\name{fcircles}
\alias{fcircles}
\title{Fresnel Circles}
\usage{
fcircles(ncircles, radius_inner = NA, radius_outer = NA, lat, lon)
}
\arguments{
\item{ncircles}{Number of equal-area concentric circles}

\item{radius_inner}{Radius of innermost circle in metres}

\item{radius_outer}{Radius of outermost circle in metres}

\item{lat}{Latitude of the centre of the Fresnel circles}

\item{lon}{Longitude of the centre of the Fresnel circles}
}
\value{
An sf dataset of Fresnel circles with a custom Azimuthal Equidistant (AEQD) CRS.
}
\description{
Function for creating Fresnel circles for mapping.
}
\examples{
fcircles(radius_inner = 125, ncircles = 8, lat = 51.51334, lon = -0.1366678)
}
