% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcircles_multi.R
\name{fcircles_multi}
\alias{fcircles_multi}
\title{Multiple Fresnel Circles}
\usage{
fcircles_multi(
  ncircles,
  radius_inner = NULL,
  radius_outer = NULL,
  geo_centres,
  id_var = NULL
)
}
\arguments{
\item{ncircles}{Number of concentric circles of equal area}

\item{radius_inner}{Radius of innermost circle in metres}

\item{radius_outer}{Radius of outermost circle in metres}

\item{geo_centres}{A spatial dataset containing the coordinates of the centres of the Fresnel circles}

\item{id_var}{Variable from geo_centres containing the location ID}
}
\value{
An sf dataset of polygons based on multiple sets of Fresnel circles across different locations.
}
\description{
Function for creating multiple sets of Fresnel circles polygons for mapping, faceted by location.
}
\examples{
# Load the sf dataset of Soho pumps
data(sohopumps)

fcircles_multi(ncircles = 2, radius_outer = 150, geo_centres = sohopumps, id_var = "Soho.Pump")
}
