% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcircles_plot.R
\name{fcircles_plot}
\alias{fcircles_plot}
\title{Map of Fresnel Circles}
\usage{
fcircles_plot(
  ncircles,
  radius_inner = NULL,
  radius_outer = NULL,
  lat = NULL,
  lon = NULL,
  geo_centre = NULL,
  geo_points
)
}
\arguments{
\item{ncircles}{Number of concentric circles of equal area including the inner circle and annuli}

\item{radius_inner}{Radius of innermost circle in metres}

\item{radius_outer}{Radius of outermost circle in metres}

\item{lat}{Latitude of the centre of the Fresnel circles}

\item{lon}{Longitude of the centre of the Fresnel circles}

\item{geo_centre}{A spatial dataset containing the coordinates of the centre of the Fresnel circles}

\item{geo_points}{A spatial dataset of points}
}
\value{
A map of Fresnel circles with overlaying points and visualised using tmap.
}
\description{
Function for plotting Fresnel circles with overlaying points. This map provides a basis for producing the Fresnel Map, which aggregates the points within the equal-area Fresnel circles
}
\examples{
# Load the sf datasets of cholera deaths and Soho pumps
data(cholera_deaths, soho_pumps)

# Filter the Broad Street Pump from the Soho pumps dataset
bstreet_pump = soho_pumps \%>\% filter(soho.pump == "Broad Street")

# Map of the Fresnel circles
fcircles_plot(radius_inner = 125, ncircles = 8, geo_centre = bstreet_pump, geo_points = cholera_deaths)
}
