% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcircles_radii.R
\name{fcircles_radii}
\alias{fcircles_radii}
\title{Fresnel Circles Radii}
\usage{
fcircles_radii(ncircles, radius_inner = NULL, radius_outer = NULL)
}
\arguments{
\item{ncircles}{Number of concentric circular zones of equal area (i.e. Fresnel circles) including the inner circle and annuli}

\item{radius_inner}{Radius of innermost Fresnel circle in metres}

\item{radius_outer}{Radius of outermost Fresnel circle in metres}
}
\value{
A dataset of Fresnel circles radii.
}
\description{
Function for creating a dataset of the Fresnel circles radii and zones.
}
\examples{
# A dataset of the Fresnel circles radii and zones
fcircles_radii(ncircles = 7, radius_outer = 300)
}
