% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmap_data.R
\name{fmap_data}
\alias{fmap_data}
\title{Polygonal Data from the Fresnel Map}
\usage{
fmap_data(
  ncircles,
  radius_inner = NA,
  radius_outer = NA,
  lat,
  lon,
  geo_points,
  sum = NA,
  mean = NA,
  median = NA
)
}
\arguments{
\item{ncircles}{Number of equal-area concentric circles}

\item{radius_inner}{Radius of innermost circle in metres}

\item{radius_outer}{Radius of outermost circle in metres}

\item{lat}{Latitude of the centre of the Fresnel Map}

\item{lon}{Longitude of the centre of the Fresnel Map}

\item{geo_points}{A spatial dataset of points to aggregate}

\item{sum}{Variable from geo_points for calculating sum}

\item{mean}{Variable from geo_pointsfor calculating mean}

\item{median}{Variable from geo_points for calculating median}
}
\value{
An sf dataset of Fresnel Map polygons with a custom Azimuthal Equidistant (AEQD) CRS based on aggregations of points-based data.
}
\description{
Function for creating Fresnel Map polygons by aggregating data to the level of equal-area concentric circles or doughnuts.
}
\examples{
data(choleradeaths)
choleradeaths = choleradeaths \%>\% st_transform(27700)
fmap_data(radius_inner = 125, ncircles = 8, lat = 51.51334, lon = -0.1366678, geo_points = choleradeaths, sum = "Cholera.Deaths")
}
