% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmap_multi.R
\name{fmap_multi}
\alias{fmap_multi}
\title{Multiple Fresnel Maps}
\usage{
fmap_multi(
  ncircles,
  radius_inner = NA,
  radius_outer = NA,
  geo_points,
  geo_centres,
  id_var = NA,
  sum = NA,
  mean = NA,
  median = NA
)
}
\arguments{
\item{ncircles}{Number of equal-area concentric circles}

\item{radius_inner}{Radius of innermost circle in metres}

\item{radius_outer}{Radius of outermost circle in metres}

\item{geo_points}{A spatial dataset of points to aggregate}

\item{geo_centres}{A spatial dataset of points to centre the Fresnel Maps}

\item{id_var}{ID column from geo_centres dataset}

\item{sum}{Variable from geo_points for calculating sum}

\item{mean}{Variable from geo_points for calculating mean}

\item{median}{Variable from geo_points for calculating median}
}
\value{
Multiple Fresnel Maps, or thematic maps with equal-area Fresnel circles, in a grid and mapped using tmap.
}
\description{
Function for creating multiple Fresnel Maps, or thematic maps that aggregate and visualise spatial data to the level of equal-area concentric circles or doughnuts, in a grid.
}
\examples{
data(choleradeaths)
data(sohopumps)
choleradeaths = choleradeaths \%>\% st_transform(27700)
sohopumps = sohopumps \%>\% st_transform(27700)
fmap_multi(ncircles = 6, radius_outer = 350, geo_points = choleradeaths, geo_centres = sohopumps, id_var = "ID", sum = "Cholera.Deaths")
}
