% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmap_stats.R
\name{fmap_stats}
\alias{fmap_stats}
\title{Summary Stats from the Fresnel Map}
\usage{
fmap_stats(
  ncircles,
  radius_inner = NA,
  radius_outer = NA,
  lat,
  lon,
  geo_points,
  sum = NA,
  mean = NA,
  median = NA
)
}
\arguments{
\item{ncircles}{Number of equal-area concentric circles}

\item{radius_inner}{Radius of innermost circle in metres}

\item{radius_outer}{Radius of outermost circle in metres}

\item{lat}{Latitude of the centre of the Fresnel Map}

\item{lon}{Longitude of the centre of the Fresnel Map}

\item{geo_points}{A spatial dataset of points to aggregate}

\item{sum}{Variable from geo_points for calculating sum}

\item{mean}{Variable from geo_points for calculating mean}

\item{median}{Variable from geo_points for calculating median}
}
\value{
Summary statistics from the Fresnel Maps including aggregated data from the Fresnel circles.
}
\description{
Function for calculating summary statistics from Fresnel Maps, or thematic maps that aggregate and visualise spatial data to the level of equal-area concentric circles or doughnuts.
}
\examples{
data(choleradeaths)
choleradeaths = choleradeaths \%>\% st_transform(27700)
fmap_stats(radius_inner = 125, ncircles = 8, lat = 51.51334, lon = -0.1366678, geo_points = choleradeaths, sum = "Cholera.Deaths")
}
