% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcircles.R
\name{fcircles}
\alias{fcircles}
\title{Fresnel Circles}
\usage{
fcircles(
  ncircles,
  radius_inner = NULL,
  radius_outer = NULL,
  lat = NULL,
  lon = NULL,
  geo_centre = NULL
)
}
\arguments{
\item{ncircles}{Number of concentric circles of equal area including the inner circle and annuli}

\item{radius_inner}{Radius of innermost circle in metres}

\item{radius_outer}{Radius of outermost circle in metres}

\item{lat}{Latitude of the centre of the Fresnel circles}

\item{lon}{Longitude of the centre of the Fresnel circles}

\item{geo_centre}{A spatial dataset containing the coordinates of the centre of the Fresnel circles}
}
\value{
An sf dataset of Fresnel circle polygons with a custom Azimuthal Equidistant (AEQD) CRS.
}
\description{
Function for creating Fresnel circles polygons for mapping.
}
\examples{
# Load the sf dataset of Soho pumps
data(soho_pumps)

# Filter the Broad Street Pump from the Soho pumps dataset
bstreet_pump = soho_pumps \%>\% filter(soho.pump == "Broad Street")

# Polygonal data from the Fresnel circles
fcircles(radius_inner = 125, ncircles = 8, geo_centre = bstreet_pump)
}
