% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcircles_multi.R
\name{fcircles_multi}
\alias{fcircles_multi}
\title{Multiple Fresnel Circles}
\usage{
fcircles_multi(
  ncircles,
  radius_inner = NULL,
  radius_outer = NULL,
  geo_centres,
  id_var = NULL,
  output = "data"
)
}
\arguments{
\item{ncircles}{Number of concentric circles of equal area}

\item{radius_inner}{Radius of innermost circle in metres}

\item{radius_outer}{Radius of outermost circle in metres}

\item{geo_centres}{A spatial dataset containing the coordinates of the centres of the Fresnel circles}

\item{id_var}{Variable from geo_centres containing the location ID}

\item{output}{Output of function. Input either 'plot' or 'data'. Defaults to 'data'}
}
\value{
An sf dataset of polygons or a simple map of the multiple Fresnel circles centred on different locations.
}
\description{
Function for plotting or deriving data from multiple Fresnel circles centred on different locations.
}
\examples{
# Load the sf dataset of Soho pumps
data(soho_pumps)

# Polygonal data from the multiple Fresnel circles
fcircles_multi(radius_outer = 105, ncircles = 2, geo_centres = soho_pumps, id_var = "soho.pump", output = "data")

# Multiple Fresnel circles
fcircles_multi(radius_outer = 105, ncircles = 2, geo_centres = soho_pumps, id_var = "soho.pump", output = "plot")
}
