% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmap_data.R
\name{fmap_data}
\alias{fmap_data}
\title{Polygonal Data from the Fresnel Map}
\usage{
fmap_data(
  ncircles,
  radius_inner = NULL,
  radius_outer = NULL,
  lat = NULL,
  lon = NULL,
  geo_centre = NULL,
  geo_points,
  sum = NULL,
  mean = NULL,
  median = NULL,
  count = F
)
}
\arguments{
\item{ncircles}{Number of equal-area concentric circles}

\item{radius_inner}{Radius of innermost circle in metres}

\item{radius_outer}{Radius of outermost circle in metres}

\item{lat}{Latitude of the centre of the Fresnel Map}

\item{lon}{Longitude of the centre of the Fresnel Map}

\item{geo_centre}{A geospatial dataset containing the coordinates of the centre of the Fresnel Map}

\item{geo_points}{A geospatial dataset of points to aggregate}

\item{sum}{Variable from geo_points for calculating sum}

\item{mean}{Variable from geo_pointsfor calculating mean}

\item{median}{Variable from geo_points for calculating median}

\item{count}{Count the number of points from geo_points. Input TRUE to count points. Defaults to FALSE}
}
\value{
An sf dataset of Fresnel Map polygons based on aggregations of points-based data with a custom Azimuthal Equidistant (AEQD) CRS.
}
\description{
Function for creating Fresnel Map polygons by aggregating data to the level of equal-area concentric circles (or annuli).
}
\examples{
# Load the sf dataset of cholera deaths
data(choleradeaths)

# Load the sf dataset of Soho pumps
data(sohopumps)

# Filter the Broad Street Pump
broadstreetpump =
  sohopumps \%>\%
  filter(Soho.Pump == "Broad Street")

fmap_data(radius_inner = 125, ncircles = 8, geo_centre = broadstreetpump, geo_points = choleradeaths, sum = "Cholera.Deaths")
}
