% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmap_data.R
\name{fmap_data}
\alias{fmap_data}
\title{Polygonal Data from the Fresnel Map}
\usage{
fmap_data(
  ncircles,
  radius_inner = NULL,
  radius_outer = NULL,
  geo_points,
  lat = NULL,
  lon = NULL,
  geo_centre = NULL,
  geo_centres = NULL,
  id_var = NULL,
  sum = NULL,
  mean = NULL,
  median = NULL,
  count = F
)
}
\arguments{
\item{ncircles}{Number of concentric circular zones of equal area (i.e. Fresnel circles) including the inner circle and annuli}

\item{radius_inner}{Radius of innermost Fresnel circle in metres}

\item{radius_outer}{Radius of outermost Fresnel circle in metres}

\item{geo_points}{A spatial dataset of points to aggregate}

\item{lat}{Latitude of the centre of the Fresnel Map}

\item{lon}{Longitude of the centre of the Fresnel Map}

\item{geo_centre}{A spatial dataset containing the coordinates of the centre of the Fresnel Map}

\item{geo_centres}{A spatial dataset containing the coordinates of the centres of each separate Fresnel Map}

\item{id_var}{Variable from geo_centres containing the location ID}

\item{sum}{Variable from geo_points for calculating sum}

\item{mean}{Variable from geo_points for calculating mean}

\item{median}{Variable from geo_points for calculating median}

\item{count}{Count the number of points from geo_points. Input TRUE to count points. Defaults to FALSE}
}
\value{
An sf dataset of Fresnel Map polygons based on aggregations of points-based data with a custom Azimuthal Equidistant (AEQD) CRS.
}
\description{
Function for creating Fresnel Map polygons by aggregating data to the level of equal-area concentric circular zones (or annuli).
}
\examples{
# Load the sf datasets of cholera deaths and Soho pumps
data(cholera_deaths, soho_pumps)

# Filter the Broad Street Pump from the Soho pumps dataset
bstreet_pump = soho_pumps \%>\% filter(soho.pump == "Broad Street")

# Polygonal data from the Fresnel Map
fmap_data(radius_inner = 125, ncircles = 8, geo_centre = bstreet_pump, geo_points = cholera_deaths, sum = "cholera.deaths")
}
