% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmap_multi.R
\name{fmap_multi}
\alias{fmap_multi}
\title{Multiple Fresnel Maps}
\usage{
fmap_multi(
  ncircles,
  radius_inner = NULL,
  radius_outer = NULL,
  geo_points,
  geo_centres,
  id_var = NULL,
  sum = NULL,
  mean = NULL,
  median = NULL,
  count = F,
  output = "plot"
)
}
\arguments{
\item{ncircles}{Number of concentric circles of equal area}

\item{radius_inner}{Radius of innermost circle in metres}

\item{radius_outer}{Radius of outermost circle in metres}

\item{geo_points}{A spatial dataset of points to aggregate}

\item{geo_centres}{A spatial dataset containing the coordinates of the centres of each Fresnel Map}

\item{id_var}{Variable from geo_centres containing the location ID}

\item{sum}{Variable from geo_points for calculating sum}

\item{mean}{Variable from geo_points for calculating mean}

\item{median}{Variable from geo_points for calculating median}

\item{count}{Count the number of points from geo_points. Input TRUE to count points. Defaults to FALSE}

\item{output}{Output of function. Input either 'plot', 'data' or 'stats'. Defaults to 'plot'}
}
\value{
A map, sf dataset or summary stats based on multiple Fresnel Maps.
}
\description{
Function for plotting or deriving data or stats from multiple Fresnel Maps, or thematic maps that visualise spatial data to the level of equal-area concentric circles (or annuli), centred on different locations.
}
\examples{
# Load the sf datasets of cholera deaths and Soho pumps
data(cholera_deaths, soho_pumps)

# Multiple Fresnel Maps
fmap_multi(radius_outer = 150, ncircles = 2, geo_points = cholera_deaths, geo_centres = soho_pumps, id_var = "soho.pump", sum = "cholera.deaths", output = "plot")

# Summary stats from the multiple Fresnel Maps
fmap_multi(radius_outer = 150, ncircles = 2, geo_points = cholera_deaths, geo_centres = soho_pumps, id_var = "soho.pump", sum = "cholera.deaths", output = "stats")

# Polygonal data from the multiple Fresnel Maps
fmap_multi(radius_outer = 150, ncircles = 2, geo_points = cholera_deaths, geo_centres = soho_pumps, id_var = "soho.pump", sum = "cholera.deaths", output = "data")
}
