% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmap_multi.R
\name{fmap_multi}
\alias{fmap_multi}
\title{Multiple Fresnel Maps}
\usage{
fmap_multi(
  ncircles,
  radius_inner = NULL,
  radius_outer = NULL,
  geo_points,
  geo_centres,
  facet = NULL,
  sum = NULL,
  mean = NULL,
  median = NULL,
  count = F
)
}
\arguments{
\item{ncircles}{Number of equal-area concentric circles}

\item{radius_inner}{Radius of innermost circle in metres}

\item{radius_outer}{Radius of outermost circle in metres}

\item{geo_points}{A geospatial dataset of points to aggregate}

\item{geo_centres}{A geospatial dataset of points across a range of locations to centre the Fresnel Maps}

\item{facet}{Variable from geo_centres for faceting the maps}

\item{sum}{Variable from geo_points for calculating sum}

\item{mean}{Variable from geo_points for calculating mean}

\item{median}{Variable from geo_points for calculating median}

\item{count}{Count the number of points. Input TRUE to count points. Defaults to FALSE}
}
\value{
Multiple Fresnel Maps in a grid and visualised using tmap.
}
\description{
Function for creating multiple Fresnel Maps, or thematic maps that visualise geospatial data to the level of equal-area concentric circles, across a range of locations.
}
\examples{
data(choleradeaths)
data(sohopumps)
choleradeaths = choleradeaths \%>\% st_transform(27700)
sohopumps = sohopumps \%>\% st_transform(27700)
fmap_multi(ncircles = 6, radius_outer = 350, geo_points = choleradeaths, geo_centres = sohopumps, facet = "Soho.Pump", sum = "Cholera.Deaths")
}
