% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmap_plot.R
\name{fmap_plot}
\alias{fmap_plot}
\title{Fresnel Map}
\usage{
fmap_plot(
  ncircles,
  radius_inner = NULL,
  radius_outer = NULL,
  lat = NULL,
  lon = NULL,
  geo_centre = NULL,
  geo_points,
  sum = NULL,
  mean = NULL,
  median = NULL,
  count = F
)
}
\arguments{
\item{ncircles}{Number of concentric circles of equal area}

\item{radius_inner}{Radius of innermost circle in metres}

\item{radius_outer}{Radius of outermost circle in metres}

\item{lat}{Latitude of the centre of the Fresnel Map}

\item{lon}{Longitude of the centre of the Fresnel Map}

\item{geo_centre}{A spatial dataset containing the coordinates of the centre of the Fresnel Map}

\item{geo_points}{A spatial dataset of points to aggregate}

\item{sum}{Variable from geo_points for calculating sum}

\item{mean}{Variable from geo_points for calculating mean}

\item{median}{Variable from geo_points for calculating median}

\item{count}{Count the number of points from geo_points. Input TRUE to count points. Defaults to FALSE}
}
\value{
A Fresnel Map based on aggregations of points-based data and visualised using tmap.
}
\description{
Function for plotting Fresnel Maps, or thematic maps that visualise geospatial data to the level of equal-area concentric circles (or annuli).
}
\examples{
# Load the sf datasets of cholera deaths and Soho pumps
data(choleradeaths, sohopumps)

# Filter the Broad Street Pump from the Soho pumps dataset
broadstreetpump = sohopumps \%>\% filter(Soho.Pump == "Broad Street")

# The Fresnel Map
fmap_plot(radius_inner = 125, ncircles = 8, geo_centre = broadstreetpump, geo_points = choleradeaths, sum = "Cholera.Deaths")
}
