% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmap_stats.R
\name{fmap_stats}
\alias{fmap_stats}
\title{Summary Stats from the Fresnel Map}
\usage{
fmap_stats(
  ncircles,
  radius_inner = NULL,
  radius_outer = NULL,
  lat = NULL,
  lon = NULL,
  geo_centre = NULL,
  geo_points,
  sum = NULL,
  mean = NULL,
  median = NULL,
  count = F
)
}
\arguments{
\item{ncircles}{Number of concentric circles of equal area including the inner circle and annuli}

\item{radius_inner}{Radius of innermost circle in metres}

\item{radius_outer}{Radius of outermost circle in metres}

\item{lat}{Latitude of the centre of the Fresnel Map}

\item{lon}{Longitude of the centre of the Fresnel Map}

\item{geo_centre}{A spatial dataset containing the coordinates of the centre of the Fresnel Map}

\item{geo_points}{A spatial dataset of points to aggregate}

\item{sum}{Variable from geo_points for calculating sum}

\item{mean}{Variable from geo_points for calculating mean}

\item{median}{Variable from geo_points for calculating median}

\item{count}{Count the number of points from geo_points. Input TRUE to count points. Defaults to FALSE}
}
\value{
Summary statistics from the Fresnel Map based on aggregated data from the Fresnel circles.
}
\description{
Function for calculating summary statistics from Fresnel Maps, or thematic maps that visualise spatial data to the level of equal-area concentric circles (or annuli).
}
\examples{
# Load the sf datasets of cholera deaths and Soho pumps
data(cholera_deaths, soho_pumps)

# Filter the Broad Street Pump from the Soho pumps dataset
bstreet_pump = soho_pumps \%>\% filter(soho.pump == "Broad Street")

# Summary stats from the Fresnel Map
fmap_stats(radius_inner = 125, ncircles = 8, geo_centre = bstreet_pump, geo_points = cholera_deaths, sum = "cholera.deaths")
}
